+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Fixes
 
-- Exit gracefully if source map file for collecting modules doesn't exist ([#2606](https://github.com/getsentry/sentry-react-native/pull/2606))
+- Android builds without ext config, auto create assets dir for modules ([#2652](https://github.com/getsentry/sentry-react-native/pull/2652))
+- Exit gracefully if source map file for collecting modules doesn't exist ([#2653](https://github.com/getsentry/sentry-react-native/pull/2653))
 
 ## 4.10.0