+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Fixes
 
+- Bump Wizard from v1.2.17 to v1.4.0 ([#2645](https://github.com/getsentry/sentry-react-native/pull/2645))
+  - [changelog](https://github.com/getsentry/sentry-wizard/blob/master/CHANGELOG.md#140)
+  - [diff](https://github.com/getsentry/sentry-wizard/compare/v1.2.17...v1.4.0)
 - Android builds without ext config, auto create assets dir for modules ([#2652](https://github.com/getsentry/sentry-react-native/pull/2652))
 
 ## 4.10.0