+++ CHANGELOG.md
@@ -9,6 +9,12 @@
   - [diff](https://github.com/getsentry/sentry-wizard/compare/v1.2.17...v1.4.0)
 - Android builds without ext config, auto create assets dir for modules ([#2652](https://github.com/getsentry/sentry-react-native/pull/2652))
 
+### Dependencies
+
+- Bump Android SDK from v6.8.0 to v6.9.1 ([#2653](https://github.com/getsentry/sentry-react-native/pull/2653))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#691)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.8.0...6.9.1)
+
 ## 4.10.0
 
 ### Features